Update docker-compose.yml
Remove a line that was unnecessary in the current implementation. I had port 8080 of the Wordpress instance forwarded to the host, as I previously used it to connect to the instance by forwarding the port through SSH. This was before I implemented SSL, so I didn't want to log in unencrypted over the web.